Support JUnit reports generated by Gradle and failsafe plugin
This action is currently made for the surefire maven plugin by setting
`/target/surefire-reports/*.xml` as the default glob to JUnit XML files.

However, the documentation of the plugin (
https://maven.apache.org/surefire/maven-surefire-report-plugin/) state
that XML report file are named `TEST-*.xml`.

Using this kind of default glob makes the action works seamlessly with
the maven failsafe plugin (see
https://maven.apache.org/surefire/maven-failsafe-plugin/) and Gradle.

Moreover having the `**` pattern allows to detect tests in submodules or
without the standard archetype of maven java project.