Improve test report with title and failure description
Before, without title, the file path were used as annotation title. But,
this information is already visible at the top of the annotation.

To maximize the amount of information, this define the test case name as
title. Also, the failure message (generally the stack trace) is provided
as raw_details of the annotation. Note that this raw_details is folded
by GitHub, so it won't reduce the readability of annotations.